+++ src/rules/__tests__/rules.test.ts
@@ -1,6 +1,5 @@
 import bestPractices = require('../best-practices');
 import errors = require('../errors');
-import style = require('../style');
 import strict = require('../strict');
 import variables = require('../variables');
 
@@ -9,7 +8,6 @@ describe('Rules', () => {
     const files = {
       'best-practices.js': bestPractices,
       'errors.js': errors,
-      'style.js': style,
       'strict.js': strict,
       'variables.js': variables
     };
@@ -20,14 +18,7 @@ describe('Rules', () => {
       Object.keys(files[file].rules).forEach(function (ruleName) {
         if (ruleName in allRules) {
           throw new Error(
-            'Found ' +
-              ruleName +
-              ' in ' +
-              file +
-              ', but ' +
-              ruleName +
-              ' already exists in ' +
-              allRules[ruleName]
+            `Found ${ruleName} in ${file}, but ${ruleName} already exists in ${allRules[ruleName]}`
           );
         }